chore: move block switcher context into page component
