--- client/src/amplify-ui/block-switcher/block-switcher.types.ts
@@ -1,7 +1,0 @@
-export type SelectedTabHeadings = string[];
-export type SetNewSelectedTabHeadings = (tabHeading: string) => void;
-
-export interface BlockSwitcherContext {
-  selectedTabHeadings: SelectedTabHeadings;
-  setNewSelectedTabHeadings: SetNewSelectedTabHeadings;
-}
