+++ client/src/docs-ui/page/page.context.tsx
@@ -3,6 +3,6 @@ import {h} from "@stencil/core";
 import {PageContext} from "./page.types";
 
 export const pageContext = createProviderConsumer<PageContext>(
-  {},
+  {} as PageContext,
   (subscribe, child) => <context-consumer {...{subscribe}} renderer={child} />,
 );
